[InlineOrder] Plugin Inline Order
Adds the ability to load a plugin to control the inline order.
This allows developing and distributing inlining heuristics
outside of tree. And together with the inline advisor plugins
allows for fine grained control of the inliner without modifying
LLVM itself.

The PluginInlineOrderAnalysis class serves as the entry point
for dynamic inline order heuristics. Plugins must register
instances of this class to provide their own InlineOrder.

Differential Revision: https://reviews.llvm.org/D140637